feat: security dashboard - handled review comments
