+++ cmd/collectors/rest/plugins/snapmirror/snapmirror.go
@@ -4,13 +4,11 @@
 package snapmirror
 
 import (
-	"encoding/json"
 	"github.com/tidwall/gjson"
 	"goharvest2/cmd/collectors/rest/plugins"
 	"goharvest2/cmd/poller/plugin"
 	"goharvest2/cmd/tools/rest"
 	"goharvest2/pkg/conf"
-	"goharvest2/pkg/errors"
 	"goharvest2/pkg/matrix"
 	"time"
 )
@@ -75,44 +73,19 @@ func (my *SnapMirror) Run(data *matrix.Matrix) ([]*matrix.Matrix, error) {
 
 func (my *SnapMirror) updateNodeCache() error {
 	var (
-		records []interface{}
-		content []byte
-		err     error
+		result []gjson.Result
+		err    error
 	)
 
 	// Clean svmVolToNode map
 	my.svmVolToNode = make(map[string]string)
-
 	href := rest.BuildHref("", "node", nil, "", "", "", "", my.query)
 
-	err = rest.FetchData(my.client, href, &records)
-	if err != nil {
-		my.Logger.Error().Stack().Err(err).Str("href", href).Msg("Failed to fetch data")
+	if result, err = plugins.InvokeRestCall(my.client, my.query, href, my.Logger); err != nil {
 		return err
 	}
 
-	all := rest.Pagination{
-		Records:    records,
-		NumRecords: len(records),
-	}
-
-	content, err = json.Marshal(all)
-	if err != nil {
-		my.Logger.Error().Err(err).Str("ApiPath", my.query).Msg("Unable to marshal rest pagination")
-	}
-
-	if !gjson.ValidBytes(content) {
-		my.Logger.Error().Err(err).Str("Api", my.query).Msg("Invalid json")
-		return errors.New(errors.API_RESPONSE, "Invalid json")
-	}
-
-	results := gjson.GetManyBytes(content, "num_records", "records")
-	numRecords := results[0]
-	if numRecords.Int() == 0 {
-		return errors.New(errors.ERR_NO_INSTANCE, "no "+my.query+" instances on cluster")
-	}
-
-	for _, volume := range results[1].Array() {
+	for _, volume := range result {
 		volumeName := volume.Get("volume").String()
 		vserverName := volume.Get("vserver").String()
 		nodeName := volume.Get("node").String()
