+++ cmd/collectors/rest/plugins/volume/volume.go
@@ -5,14 +5,12 @@
 package volume
 
 import (
-	"encoding/json"
 	"github.com/tidwall/gjson"
 	"goharvest2/cmd/collectors/rest/plugins"
 	"goharvest2/cmd/poller/plugin"
 	"goharvest2/cmd/tools/rest"
 	"goharvest2/pkg/conf"
 	"goharvest2/pkg/dict"
-	"goharvest2/pkg/errors"
 	"goharvest2/pkg/matrix"
 	"goharvest2/pkg/tree/node"
 	"strconv"
@@ -143,46 +141,21 @@ func (my *Volume) getDataInterval(param *node.Node, defaultInterval time.Duratio
 
 func (my *Volume) GetSnapMirrors() (map[string][]*matrix.Instance, map[string]*matrix.Instance, error) {
 	var (
-		records []interface{}
-		content []byte
-		err     error
+		result []gjson.Result
+		err    error
 	)
 
 	smSourceMap := make(map[string][]*matrix.Instance)
 	smDestinationMap := make(map[string]*matrix.Instance)
 
 	snapmirrorData := matrix.New(my.Parent+".SnapMirror", "sm", "sm")
-
 	href := rest.BuildHref("", strings.Join(my.snapmirrorFields, ","), nil, "", "", "", "", my.query)
 
-	err = rest.FetchData(my.client, href, &records)
-	if err != nil {
-		my.Logger.Error().Stack().Err(err).Str("href", href).Msg("Failed to fetch data")
+	if result, err = plugins.InvokeRestCall(my.client, my.query, href, my.Logger); err != nil {
 		return nil, nil, err
 	}
 
-	all := rest.Pagination{
-		Records:    records,
-		NumRecords: len(records),
-	}
-
-	content, err = json.Marshal(all)
-	if err != nil {
-		my.Logger.Error().Err(err).Str("ApiPath", my.query).Msg("Unable to marshal rest pagination")
-	}
-
-	if !gjson.ValidBytes(content) {
-		my.Logger.Error().Err(err).Str("Api", my.query).Msg("Invalid json")
-		return nil, nil, errors.New(errors.API_RESPONSE, "Invalid json")
-	}
-
-	results := gjson.GetManyBytes(content, "num_records", "records")
-	numRecords := results[0]
-	if numRecords.Int() == 0 {
-		return nil, nil, errors.New(errors.ERR_NO_INSTANCE, "no "+my.query+" instances on cluster")
-	}
-
-	for _, snapMirror := range results[1].Array() {
+	for _, snapMirror := range result {
 
 		relationshipId := snapMirror.Get("relationship_id").String()
 		groupType := snapMirror.Get("relationship_group_type").String()
@@ -326,44 +299,18 @@ func (my *Volume) updateVolumeLabels(data *matrix.Matrix) {
 
 func (my *Volume) getDiskData() ([]gjson.Result, error) {
 	var (
-		records []interface{}
-		content []byte
-		err     error
+		result []gjson.Result
+		err    error
 	)
 
 	diskFields := []string{"aggregates.name", "aggregates.uuid"}
 	query := "api/storage/disks"
-
 	href := rest.BuildHref("", strings.Join(diskFields, ","), []string{"protection_mode=!data|full"}, "", "", "", "", query)
 
-	err = rest.FetchData(my.client, href, &records)
-	if err != nil {
-		my.Logger.Error().Stack().Err(err).Str("href", href).Msg("Failed to fetch data")
+	if result, err = plugins.InvokeRestCall(my.client, query, href, my.Logger); err != nil {
 		return nil, err
 	}
-
-	all := rest.Pagination{
-		Records:    records,
-		NumRecords: len(records),
-	}
-
-	content, err = json.Marshal(all)
-	if err != nil {
-		my.Logger.Error().Err(err).Str("ApiPath", query).Msg("Unable to marshal rest pagination")
-		return nil, err
-	}
-
-	if !gjson.ValidBytes(content) {
-		my.Logger.Error().Err(err).Str("Api", query).Msg("Invalid json")
-		return nil, errors.New(errors.API_RESPONSE, "Invalid json")
-	}
-
-	results := gjson.GetManyBytes(content, "num_records", "records")
-	numRecords := results[0]
-	if numRecords.Int() == 0 {
-		return nil, errors.New(errors.ERR_NO_INSTANCE, "no "+my.query+" instances on cluster")
-	}
-	return results[1].Array(), nil
+	return result, nil
 }
 
 func (my *Volume) updateAggrMap(disks []gjson.Result) {
